VCL canvas: fix clipping of gradients (related: tdf#115142)
When using the presenter console, the secondary screen has
a small view of the current slide, which is called a "Pane" in sdext.

Clipping for this is done by sd::presenter::PresenterCanvas, which wraps
the full-screen VCL canvas and has a MergeViewState() to intersect
the clip region.

The VCL canvas has a bug in its gradientFill() implemenation where it
overrides the existing clip on the OutputDevice instead of intersecting
it.

This properly fixes the bug that was wrongly and badly attempted
to be fixed by commit 90dfa877b79d4c24d8c306b3502525e3857cbd70.

Change-Id: Id02678c2f134fe7b190866d6a6fff5b2e5e3aaff