Fix external/libassuan Linux RPATH
In other external projects using libtool, we fix that by patching configure,
resetting hardcode_libdir_flag_spec[_CXX] at the end of the linux*) case block
that sets the Linux-specific value.  But here we run autoreconf in
ExternalProject_libassuan, so that patch in configure would be overwritten.  The
relevant code in configure comes from autoconf boilerplate, so we cannot just do
the same patch in configure.ac.  But we can reset hardcode_libdir_flag_spec
sufficiently late in configure.ac so that things still work as intended.

Change-Id: Ic6c6123bcfe8bc2dac87812f919842519374abaa
Reviewed-on: https://gerrit.libreoffice.org/50959
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>