iOS, moved privateSnippetExecutor from asm to C function.
move asm code to C as first step towards reducing the asm code.
Since iOS does not permit java/python or anything else than the
compiled C++ code, the throw should be done simpler.

Apart from that this is the first step in solving a stack
corruption problem in the throw code

Change-Id: I4f3d3a3ba3f55fb46131d9a8eeb0deebf179d95f