change recursive ImageURL prop. search to use XGraphic
In two cases we need to traverse and gather all ImageURL properties
and get the URL string and store the graphic content to a storage
(like we do in xmloff filter). ImageURL property can now only store
external URL and Graphic stores the embedded XGraphic, so this was
changed to look into Graphic property first and then ImageURL.

We also don't gather URL sting anymore so they need to be loaded
to XGraphic when gathering them.

Change-Id: I5f3f4be2b403b9589d72b8733df0c97109f2b65d
Reviewed-on: https://gerrit.libreoffice.org/51308
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>