forcepoint #3: sw: don't do section-in-table splits when this is table
The problem with the bugdoc is that SwFrame::GetNextSctLeaf() creates a
follow frame for a section in a table, and puts it outside the table,
below the page's body frame, while moving text frames that have the
mbInfTab flag set into the follow frame.

The situation is different from forcepoint #18 though: the frame "this"
is a SwTabFrame here, and GetNextCellLeaf() *does* return a plausible
looking follow cell frame, however it's ignored later because
CanContainSplitSection(this) is false since it's a SwTabFrame.

IIRC table-in-section-in-table should not be split anyway, so
just add a check here and return early.

(regression 0640c594254d1060ec416706e54f02788b7f1aa4)

Revert "forcepoint #3"
This reverts commit 47916d6bd84af3f6ef684404ab6ffc5af57ee685
because it was just a work-around not to crash with the invalid table
frames.

Change-Id: I54692a67edade23c6b8e0a9be806728b0012a3fd
Reviewed-on: https://gerrit.libreoffice.org/51751
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>