Remove bogus assert
...that had been added with 94354a0646521ce5249e271f116d271038fe1b2a "unotools:
assert if TempFile::GetURL() fails due to missing file UCP".  But when SRCDIR is
a read-only tree, JunitTest_sc_unoapi_2 fails with that assert firing when
TempFile::GetURL is called from SfxMedium::CreateTempFile
(sfx2/source/doc/docfile.cxx), which in turn is prepared to handle an empty
return value from GetURL, and the test succeeds after removing the assert.  So
it looks like that assert was not precise enough to only fire in the scenarios
where it was intended to fire.

Change-Id: I333e9026ee14e9d1140b1b475793a5b15ededd70
Reviewed-on: https://gerrit.libreoffice.org/52085
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>