tdf#116666 Hungarian collation: casing and equality fixes
Casing fixes: “CCS” sorted as “CSCS”, not “cscs”.
“Ccs” and “CCS” are capitalized versions of the simplified
double consonant “cs”, but “CCs” is an abbreviation of words
beginning with “C” and “Cs” (similar to “AkH.”, “MHSz.”) etc.

To avoid the comparison result “equal” we set a precedence
between the simplified and compound-like long forms, too.
For example, “ésszerű” (old orthography before 2015) and
“észszerű” (not “észszerű”, “ésszerű”), or “mennyelv” and
“menynyelv” (words with different meanings) sorted as
“észszerű” and “észSzerű”, also “menynyelv” and “menyNyelv”.

Change-Id: If31c97262bc74429b514ede43a0384de80fe8ac5
Reviewed-on: https://gerrit.libreoffice.org/52194
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Andras Timar <andras.timar@collabora.com>