tdf#116265 RTF import: ignore \fi in para if it repeats invalid \fi from list
The 2nd paragraph of the bugdoc has a \fi set, but it has to be ignored.
The reason for this is that the \fi is a repeated property from the list
definition, but there \levelnumbers has invalid contents, which means
its \fi has to be ignored, including their repeated values in paragraph
direct formatting.

There was already code to insert implicit defaults as paragraph direct
formatting based on list properties, this does the opposite: remove
paragraph direct formatting based on the list state.

This requires tracking the ignored values in a separate structure, since
the normal list properties no longer contain the information we need to
be able to remove paragraph formatting. (At the moment this tracks only
\fi, but can be extended to other properties if it'll be necessary.)

Change-Id: Iff542cb76c66b7fc7ae4e07ade84701e24a74f1a
Reviewed-on: https://gerrit.libreoffice.org/52352
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>