tdf#116789 sw: fix caching of the SwXCell UNO wrapper
Commit 7c05136d49ddee6eb1976a9970be3af133e3a7c1 (unify with a template
class, 2015-03-28) extracted a common FindUnoInstanceHint from
FindXCellHint/FindXRowHint, but failed to still check for the UNO
wrapper in the condition, not the core object.

This resulted in handing out multiple SwXCell instances for the same
underlying SwTableBox, which means that if you had an UNO-level
selection listener and you wanted to know if your selection is in the
same XText as the XText of a bookmark, then that comparison (by ptr
value) failed in case the bookmark and the selection was in a table
cell.

Change-Id: I1d618cf150d515a9db51b1fe5d38cf262fd46af4
Reviewed-on: https://gerrit.libreoffice.org/52379
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>