tdf#116830 sw: handle cell background for BACKGROUND_PARA_OVER_DRAWINGS
Commit 43b8fe8adda414f8107e685add8c8dd4b19795b6 (n#778133 sw: add
BackgroundParaOverDrawings compat flag, 2012-09-12) added a layout
compat flag that is meant to make Writer behave the same as Word when it
comes to the order of various backgrounds.

The original fix was that in case there are background shapes and also
paragraph backgrounds, then this flag can opt in to have the paragraph
backgrounds on top of background shapes.

Cell backgrounds behave the same way in Word, so extend the behavior of
the compat flag to cover not only the background of text frames but also
of cell frames.

(The test could be hacked into CppunitTest_sw_uiwriter, but 1) this is
layout, not UI and 2) that one started to get rather large.)

Change-Id: Ifca1851217000f6daac50e6c9562c042aca6ff7d
Reviewed-on: https://gerrit.libreoffice.org/52469
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>