tdf#50057 sw: fix duplication of at-paragraph anchored flys
The problem is that the flys that are anchored to the first paragraph of
a SwRedline do get copied by SwRangeRedline::CopyToSection(), but they
don't get deleted by DelCopyOfSection().

Copying is enabled because the setting of the SetRedlineMove()
was made conditional in commit 65de5382a389cc7edf1cdf506da4fb43a4d33a9f
"#100619# fixed problem with hiding deleted graphics",
which isn't justified in any way and which i can't imaginge a good reason
for, given that the flys anchored in the first paragraph are all skipped
in DelFlyInRange(), so why would you want to copy them; hence this
reverts that commit.

The interesting check for the redline_fly_duplication_at_para_end_inside
case is actually the one that was added in commit
e4d835d351f839fe75eaf215307886795d3861ab

Change-Id: I96fb294a5456e7f1172a5f408ebcb21cf211c276
Reviewed-on: https://gerrit.libreoffice.org/52729
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>