tdf#117028 sw: conditionally consider fill of shape for in-textbox auto color
Commit 4e0c8e07485596804d512b8dc9d11984381f882d (tdf#116925 sw: consider
fill styles of shape for in-textbox auto color, 2018-04-10) did this
unconditionally, but there are two cases:

1) Shape (with textbox) has a fill, in that case auto text color of
textbox should consider the shape fill.

2) Shape has no fill, in that case the shape fill should not be
considered when counting the background of a transparent fly frame.

Fix 2) without breaking 1) by making the change to
SwFrame::GetBackgroundBrush() opt-in, and only use it for auto font
color, not in other cases.

Change-Id: I761a1b7d6ecf47586edbe5641ec012f22f6a53d7
Reviewed-on: https://gerrit.libreoffice.org/52941
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>