tdf#86399 Correctly set cluster start for RTL runs
A followup for the fix in 5d0862217160bec3da95b8aee2042142ddc67065 (from
2015) that didn’t work for Arabic because we were setting cluster start
for RTL runs incorrectly.

I’ve been suffering from this for so long but too lazy to debug it,
until I saw this pace of code by pure chance and it looked dubious,
finding where this flag is used and then seeing the comment referring to
tdf#86399, then WOW I know what is going on!

Change-Id: Ib7850e5a90ed3184738e995885f7f8b459ea2bb9
Reviewed-on: https://gerrit.libreoffice.org/53261
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>