svx: import processed PDF text
Some PDFs don't include spaces in the text.
Instead, they rely on the explicit positioning
of each character to render visually separated words.
Latex seems to be prone to this approach, though not
exclusively.

Luckily, PDFium does process text and inserts
"generated" spaces where necessary, which is what
we retrieve and use as the text string while importing.

Change-Id: Ic21fe6c8416ecaba66f06b6260f1d6b040ff12af