tdf#117017 Pasting into the formula bar shouldn't retain formatting
This is (kind of) regression from the previous commit
60f70b6cc2eedbf22a54a77352ced8ac44f23828 ("tdf#117017 Make modifier
keys not trigger the InputChanged handler").

Reset of formatting when pasting into the formula bar via a keyboard
shortcut, was actually a side effect of the gtk auto-accelerator
work. In particual, bibisect (of the 5.2 cycle) points to commit
3af8f4fa79447916b3af14069bc649b667f41931 ("tdf#96739: Send Ctrl-Left/
RightShift events to correct window"). The behavior was later changed
back with commit a5837d3328a5f1111b16e1cb6b89d13936525337
("Resolves: tdf#99324 let sidebar toggle auto-mnemonics on/off with
alt"), but introduced again upon request of tdf#113894, with commit
60f16d43b453cc135907c2394b53034672e4a4f9 ("tdf#113894 release ctrl of
ctrl+v in input line should strip formatting"). Now it's broken again
with the previous commit of this bug.

However, even if not retaining formatting is the desired behavior,
triggering it on a modifier key release isn't a good idea. So keep
the previous commit, and implement tdf#113894 differently. In
addition, make sure to get identical behavior regardless of the
pasting method, be it via a keyboard shortcut, a toolbar button or a
menu/context menu command. And users can still override this behavior
with the usual paste special dialog, or the toolbar dropdown.

Change-Id: Icd801d62ceeead066484256d60f9585f2e7ebaec
Reviewed-on: https://gerrit.libreoffice.org/53516
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>